# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEROAD",
        "description": "Api documentation",
        "contact": {
            "email": "peramkat@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://theme.zdassets.com/theme_assets/9115960/ef5800cc529889d180b05b57e40dd50e5c7adb73.png"
        }
    },
    "paths": {
...
        "/api/v1/travels/{travel:slug}/tours": {
            "get": {
                "tags": [
                    "Api v1 - Tours"
                ],
                "summary": "Retrieve tours for a specific travel",
                "description": "Get a list of tours for a specific travel based on the provided filters",
                "operationId": "a1084137b606dfdde9b1efe97ee047d1",
                "parameters": [
                    {
                        "name": "travel:slug",
                        "in": "path",
                        "description": "Slug of the travel to retrieve tours for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "jordan-360"
                    },
                    {
                        "name": "priceFrom",
                        "in": "query",
                        "description": "Minimum price for filtering tours",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "priceTo",
                        "in": "query",
                        "description": "Maximum price for filtering tours",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "description": "Minimum starting date for filtering tours",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "description": "Maximum ending date for filtering tours",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Field to sort the tours by",
                        "schema": {
                            "type": "string"
                        },
                        "example": "price"
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Sorting order ('asc' or 'desc')",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tours for the specified travel"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/travels/{travel:slug}/tour": {
            "post": {
                "tags": [
                    "Api v1 - Tours"
                ],
                "summary": "Create a new tour for a travel",
                "description": "Store a newly created tour for a specific travel in the database",
                "operationId": "5814af6430b23edaa8bcdfdba62bf1ea",
                "parameters": [
                    {
                        "name": "travel:slug",
                        "in": "path",
                        "description": "Slug of the travel to associate the tour with",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "varchar"
                        },
                        "example": "jordan-360"
                    }
                ],
                "requestBody": {
                    "description": "Tour details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "new tour"
                                    },
                                    "startingDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-01"
                                    },
                                    "endingDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-10"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "2345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tour created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}